requirements.txt
Django==3.1.6
djangorestframework==3.12.2
flake8==3.8.4
environ


